move test into test_csv
